refactor: refactored thread vs message conflation
