+++ core/src/inference/index.ts
@@ -1,1 +1,2 @@
-export * from './inference'
+export * from './inferenceEntity'
+export * from './inferenceInterface'
