+++ core/src/message/index.ts
@@ -1,0 +1,2 @@
+export * from './messageEntity'
+export * from './messageInterface'
