+++ core/src/message/messageEntity.ts
@@ -1,0 +1,88 @@
+import { ChatCompletionMessage, ChatCompletionRole } from '../inference'
+import { ModelInfo } from '../model'
+
+/**
+ * The `ThreadMessage` type defines the shape of a thread's message object.
+ * @stored
+ */
+export interface ThreadMessage {
+  /** Unique identifier for the message, generated by default using the ULID method. **/
+  id: string
+  /** Object name **/
+  object: string
+  /** Thread id, default is a ulid. **/
+  thread_id: string
+  /** The role of the author of this message. **/
+  assistant_id?: string
+  // TODO: comment
+  role: ChatCompletionRole
+  /** The content of this message. **/
+  content: ThreadContent[]
+  /** The status of this message. **/
+  status: MessageStatus
+  /** The timestamp indicating when this message was created. Represented in Unix time. **/
+  created: number
+  /** The timestamp indicating when this message was updated. Represented in Unix time. **/
+  updated: number
+  /** The additional metadata of this message. **/
+  metadata?: Record<string, unknown>
+}
+
+/**
+ * The `MessageRequest` type defines the shape of a new message request object.
+ * @data_transfer_object
+ */
+export interface MessageRequest {
+  id?: string
+
+  /** The thread id of the message request. **/
+  threadId: string
+
+  /**
+   * The assistant id of the message request.
+   */
+  assistantId?: string
+
+  /** Messages for constructing a chat completion request **/
+  messages?: ChatCompletionMessage[]
+
+  /** Settings for constructing a chat completion request **/
+  model?: ModelInfo
+}
+
+/**
+ * The `ThreadContent` type defines the shape of a message's content object
+ * @data_transfer_object
+ */
+export interface ThreadContent {
+  type: ContentType
+  text: ContentValue
+}
+
+/**
+ * The status of the message.
+ * @data_transfer_object
+ */
+export enum MessageStatus {
+  /** Message is fully loaded. **/
+  Ready = 'ready',
+  /** Message is not fully loaded. **/
+  Pending = 'pending',
+}
+
+/**
+ * The content type of the message.
+ */
+export enum ContentType {
+  Text = 'text',
+  Image = 'image',
+}
+
+/**
+ * The `ContentValue` type defines the shape of a content value object
+ * @data_transfer_object
+ */
+export interface ContentValue {
+  value: string
+  annotations: string[]
+}
