+++ core/src/monitoring/index.ts
@@ -1,1 +1,1 @@
-export * from './monitoring'
+export * from './monitoringInterface'
